@@ -8,7 +8,7 @@ import (
 type Addition struct {
 	Cookie       string  `json:"cookie" type:"text" help:"one of QR code token and cookie required"`
 	QRCodeToken  string  `json:"qrcode_token" type:"text" help:"one of QR code token and cookie required"`
-	QRCodeSource string  `json:"qrcode_source" type:"select" options:"web,android,ios,linux,mac,windows,tv" default:"linux" help:"select the QR code device, default linux"`
+	QRCodeSource string  `json:"qrcode_source" type:"select" options:"web,android,ios,tv,alipaymini,wechatmini,qandroid" default:"linux" help:"select the QR code device, default linux"`
 	PageSize     int64   `json:"page_size" type:"number" default:"56" help:"list api per page size of 115 driver"`
 	LimitRate    float64 `json:"limit_rate" type:"number" default:"2" help:"limit all api request rate (1r/[limit_rate]s)"`
 	driver.RootID
@@ -17,9 +17,9 @@ type Addition struct {
 var config = driver.Config{
 	Name:        "115 Cloud",
 	DefaultRoot: "0",
-	//OnlyProxy:   true,
-	//OnlyLocal:         true,
-	//NoOverwriteUpload: true,
+	// OnlyProxy:   true,
+	// OnlyLocal:         true,
+	// NoOverwriteUpload: true,
 }
 
 func init() {
