@@ -26,7 +26,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-var UserAgent = driver115.UA115Desktop
+var UserAgent = driver115.UA115Browser
 
 func (d *Pan115) login() error {
 	var err error
@@ -38,17 +38,17 @@ func (d *Pan115) login() error {
 	}
 	d.client = driver115.New(opts...)
 	cr := &driver115.Credential{}
-	if d.Addition.QRCodeToken != "" {
+	if d.QRCodeToken != "" {
 		s := &driver115.QRCodeSession{
-			UID: d.Addition.QRCodeToken,
+			UID: d.QRCodeToken,
 		}
 		if cr, err = d.client.QRCodeLoginWithApp(s, driver115.LoginApp(d.QRCodeSource)); err != nil {
 			return errors.Wrap(err, "failed to login by qrcode")
 		}
-		d.Addition.Cookie = fmt.Sprintf("UID=%s;CID=%s;SEID=%s", cr.UID, cr.CID, cr.SEID)
-		d.Addition.QRCodeToken = ""
-	} else if d.Addition.Cookie != "" {
-		if err = cr.FromCookie(d.Addition.Cookie); err != nil {
+		d.Cookie = fmt.Sprintf("UID=%s;CID=%s;SEID=%s", cr.UID, cr.CID, cr.SEID)
+		d.QRCodeToken = ""
+	} else if d.Cookie != "" {
+		if err = cr.FromCookie(d.Cookie); err != nil {
 			return errors.Wrap(err, "failed to login by cookies")
 		}
 		d.client.ImportCredential(cr)
@@ -74,8 +74,22 @@ func (d *Pan115) getFiles(fileId string) ([]FileObj, error) {
 }
 
 const (
-	appVer = "2.0.3.6"
+	appVer = "27.0.3.7"
 )
+
+func (c *Pan115) getAppVer() string {
+	// todo add some cache？
+	vers, err := c.client.GetAppVersion()
+	if err != nil {
+		return appVer
+	}
+	for _, ver := range vers {
+		if ver.AppName == "win" {
+			return ver.Version
+		}
+	}
+	return appVer
+}
 
 func (c *Pan115) DownloadWithUA(pickCode, ua string) (*driver115.DownloadInfo, error) {
 	key := crypto.GenerateKey()
@@ -151,7 +165,7 @@ func (d *Pan115) rapidUpload(fileSize int64, fileName, dirID, preID, fileID stri
 	userID := strconv.FormatInt(d.client.UserID, 10)
 	form := url.Values{}
 	form.Set("appid", "0")
-	form.Set("appversion", appVer)
+	form.Set("appversion", d.getAppVer())
 	form.Set("userid", userID)
 	form.Set("filename", fileName)
 	form.Set("filesize", fileSizeStr)
@@ -161,7 +175,7 @@ func (d *Pan115) rapidUpload(fileSize int64, fileName, dirID, preID, fileID stri
 
 	signKey, signVal := "", ""
 	for retry := true; retry; {
-		t := driver115.Now()
+		t := driver115.NowMilli()
 
 		if encodedToken, err = ecdhCipher.EncodeToken(t.ToInt64()); err != nil {
 			return nil, err
@@ -225,6 +239,9 @@ func UploadDigestRange(stream model.FileStreamer, rangeSpec string) (result stri
 
 	length := end - start + 1
 	reader, err := stream.RangeRead(http_range.Range{Start: start, Length: length})
+	if err != nil {
+		return "", err
+	}
 	hashStr, err := utils.HashReader(utils.SHA1, reader)
 	if err != nil {
 		return "", err
@@ -370,11 +387,13 @@ LOOP:
 	}
 	return d.checkUploadStatus(dirID, params.SHA1)
 }
+
 func chunksProducer(ch chan oss.FileChunk, chunks []oss.FileChunk) {
 	for _, chunk := range chunks {
 		ch <- chunk
 	}
 }
+
 func (d *Pan115) checkUploadStatus(dirID, sha1 string) error {
 	// 验证上传是否成功
 	req := d.client.NewRequest().ForceContentType("application/json;charset=UTF-8")
@@ -431,8 +450,8 @@ func SplitFileByPartNum(fileSize int64, chunkNum int) ([]oss.FileChunk, error) {
 	}
 
 	var chunks []oss.FileChunk
-	var chunk = oss.FileChunk{}
-	var chunkN = (int64)(chunkNum)
+	chunk := oss.FileChunk{}
+	chunkN := (int64)(chunkNum)
 	for i := int64(0); i < chunkN; i++ {
 		chunk.Number = int(i + 1)
 		chunk.Offset = i * (fileSize / chunkN)
@@ -454,13 +473,13 @@ func SplitFileByPartSize(fileSize int64, chunkSize int64) ([]oss.FileChunk, erro
 		return nil, errors.New("chunkSize invalid")
 	}
 
-	var chunkN = fileSize / chunkSize
+	chunkN := fileSize / chunkSize
 	if chunkN >= 10000 {
 		return nil, errors.New("Too many parts, please increase part size")
 	}
 
 	var chunks []oss.FileChunk
-	var chunk = oss.FileChunk{}
+	chunk := oss.FileChunk{}
 	for i := int64(0); i < chunkN; i++ {
 		chunk.Number = int(i + 1)
 		chunk.Offset = i * chunkSize
