@@ -1,6 +1,7 @@
 package _123Share
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"hash/crc32"
@@ -80,20 +81,19 @@ func (d *Pan123Share) request(url string, method string, callback base.ReqCallba
 	return body, nil
 }
 
-func (d *Pan123Share) getFiles(parentId string) ([]File, error) {
+func (d *Pan123Share) getFiles(ctx context.Context, parentId string) ([]File, error) {
 	page := 1
 	res := make([]File, 0)
 	for {
-		if !d.APIRateLimit(FileList) {
-			time.Sleep(time.Millisecond * 200)
-			continue
+		if err := d.APIRateLimit(ctx, FileList); err != nil {
+			return nil, err
 		}
 		var resp Files
 		query := map[string]string{
 			"limit":          "100",
 			"next":           "0",
-			"orderBy":        d.OrderBy,
-			"orderDirection": d.OrderDirection,
+			"orderBy":        "file_id",
+			"orderDirection": "desc",
 			"parentFileId":   parentId,
 			"Page":           strconv.Itoa(page),
 			"shareKey":       d.ShareKey,
