@@ -137,13 +137,19 @@ func (d *BaiduPhoto) Link(ctx context.Context, file model.Obj, args model.LinkAr
 	case *File:
 		return d.linkFile(ctx, file, args)
 	case *AlbumFile:
-		f, err := d.CopyAlbumFile(ctx, file)
-		if err != nil {
-			return nil, err
+		// 处理共享相册
+		if d.Uk != file.Uk {
+			// 有概率无法获取到链接
+			return d.linkAlbum(ctx, file, args)
+
+			// 接口被限制，只能使用cookie
+			// f, err := d.CopyAlbumFile(ctx, file)
+			// if err != nil {
+			// 	return nil, err
+			// }
+			// return d.linkFile(ctx, f, args)
 		}
-		return d.linkFile(ctx, f, args)
-		// 有概率无法获取到链接
-		//return d.linkAlbum(ctx, file, args)
+		return d.linkFile(ctx, &file.File, args)
 	}
 	return nil, errs.NotFile
 }
