@@ -9,13 +9,12 @@ type Addition struct {
 	// Usually one of two
 	driver.RootPath
 	// define other
-	RefreshToken         string `json:"refresh_token" required:"true" help:"login type is refresh_token,this is required"`
-	UploadThread         string `json:"upload_thread" default:"3" help:"1 <= thread <= 32"`
-	CustomUploadPartSize string `json:"custom_upload_part_size" default:"0" help:"0 for auto"`
+	RefreshToken string `json:"refresh_token" required:"true" help:"login type is refresh_token,this is required"`
+	UploadThread string `json:"upload_thread" default:"3" help:"1 <= thread <= 32"`
 
-	AppID      string `json:"app_id" required:"true" default:"devDebugger/1.0"`
+	AppID      string `json:"app_id" required:"true" default:"alist/10001"`
 	AppVersion string `json:"app_version" required:"true" default:"1.0.0"`
-	AppSecret  string `json:"app_secret" required:"true" default:"Nkx3Y2xvZ2luLmNu"`
+	AppSecret  string `json:"app_secret" required:"true" default:"bR4SJwOkvnG5WvVJ"`
 }
 
 var config = driver.Config{
