@@ -4,6 +4,7 @@ import (
 	"github.com/alist-org/alist/v3/internal/model"
 	"github.com/alist-org/alist/v3/pkg/utils"
 	"github.com/city404/v6-public-rpc-proto/go/v6/common"
+	pubUserFile "github.com/city404/v6-public-rpc-proto/go/v6/userfile"
 	"google.golang.org/grpc"
 	"time"
 )
@@ -50,34 +51,7 @@ type FilesList struct {
 
 var _ model.Obj = (*Files)(nil)
 
-type Files struct {
-	Identity        string `json:"identity,omitempty"`
-	Parent          string `json:"parent,omitempty"`
-	Name            string `json:"name,omitempty"`
-	Path            string `json:"path,omitempty"`
-	MimeType        string `json:"mime_type,omitempty"`
-	Size            int64  `json:"size,omitempty"`
-	Type            int64  `json:"type,omitempty"`
-	CreateTs        int64  `json:"create_ts,omitempty"`
-	UpdateTs        int64  `json:"update_ts,omitempty"`
-	DeleteTs        int64  `json:"delete_ts,omitempty"`
-	Deleted         bool   `json:"deleted,omitempty"`
-	Dir             bool   `json:"dir,omitempty"`
-	Hidden          bool   `json:"hidden,omitempty"`
-	Locked          bool   `json:"locked,omitempty"`
-	Shared          bool   `json:"shared,omitempty"`
-	Starred         bool   `json:"starred,omitempty"`
-	Trashed         bool   `json:"trashed,omitempty"`
-	LockedAt        int64  `json:"locked_at,omitempty"`
-	LockedBy        string `json:"locked_by,omitempty"`
-	SharedAt        int64  `json:"shared_at,omitempty"`
-	Flag            int64  `json:"flag,omitempty"`
-	Unique          string `json:"unique,omitempty"`
-	ContentIdentity string `json:"content_identity,omitempty"`
-	Label           int64  `json:"label,omitempty"`
-	StoreType       int64  `json:"store_type,omitempty"`
-	Version         int64  `json:"version,omitempty"`
-}
+type Files pubUserFile.File
 
 func (f *Files) GetSize() int64 {
 	return f.Size
@@ -112,20 +86,6 @@ func (f *Files) GetID() string {
 
 func (f *Files) GetPath() string {
 	return f.Path
-}
-
-func fileToObj(f Files) *model.ObjThumb {
-	return &model.ObjThumb{
-		Object: model.Object{
-			ID:       f.Identity,
-			Path:     f.Path,
-			Name:     f.Name,
-			Size:     f.Size,
-			Modified: f.ModTime(),
-			Ctime:    f.CreateTime(),
-			IsFolder: f.IsDir(),
-		},
-	}
 }
 
 type SteamFile struct {
