@@ -7,11 +7,16 @@ import (
 
 type Addition struct {
 	driver.RootID
-	Username         string `json:"username" required:"true"`
-	Password         string `json:"password" required:"true"`
-	ClientID         string `json:"client_id" required:"true" default:"YNxT9w7GMdWvEOKa"`
-	ClientSecret     string `json:"client_secret" required:"true" default:"dbw2OtmVEeuUvIptb1Coyg"`
-	DisableMediaLink bool   `json:"disable_media_link"`
+	Username                string `json:"username" required:"true"`
+	Password                string `json:"password" required:"true"`
+	Platform                string `json:"platform" required:"true" type:"select" options:"android,web,pc"`
+	RefreshToken            string `json:"refresh_token" required:"true" default:""`
+	RefreshTokenMethod      string `json:"refresh_token_method" required:"true" type:"select" options:"oauth2,http"`
+	CaptchaToken            string `json:"captcha_token" default:""`
+	DeviceID                string `json:"device_id"  required:"false" default:""`
+	DisableMediaLink        bool   `json:"disable_media_link" default:"true"`
+	UseLowLatencyAddress    bool   `json:"use_low_latency_address" default:"false"`
+	CustomLowLatencyAddress string `json:"custom_low_latency_address" default:""`
 }
 
 var config = driver.Config{
