@@ -1,20 +1,16 @@
 package pikpak_share
 
 import (
+	"fmt"
 	"strconv"
 	"time"
 
 	"github.com/alist-org/alist/v3/internal/model"
 )
 
-type RespErr struct {
-	ErrorCode int    `json:"error_code"`
-	Error     string `json:"error"`
-}
-
 type ShareResp struct {
-	ShareStatus		string `json:"share_status"`
-	ShareStatusText	string `json:"share_status_text"`
+	ShareStatus     string `json:"share_status"`
+	ShareStatusText string `json:"share_status_text"`
 	FileInfo        File   `json:"file_info"`
 	Files           []File `json:"files"`
 	NextPageToken   string `json:"next_page_token"`
@@ -77,4 +73,33 @@ type Media struct {
 	ResolutionName string        `json:"resolution_name"`
 	IsVisible      bool          `json:"is_visible"`
 	Category       string        `json:"category"`
+}
+
+type CaptchaTokenRequest struct {
+	Action       string            `json:"action"`
+	CaptchaToken string            `json:"captcha_token"`
+	ClientID     string            `json:"client_id"`
+	DeviceID     string            `json:"device_id"`
+	Meta         map[string]string `json:"meta"`
+	RedirectUri  string            `json:"redirect_uri"`
+}
+
+type CaptchaTokenResponse struct {
+	CaptchaToken string `json:"captcha_token"`
+	ExpiresIn    int64  `json:"expires_in"`
+	Url          string `json:"url"`
+}
+
+type ErrResp struct {
+	ErrorCode        int64  `json:"error_code"`
+	ErrorMsg         string `json:"error"`
+	ErrorDescription string `json:"error_description"`
+}
+
+func (e *ErrResp) IsError() bool {
+	return e.ErrorCode != 0 || e.ErrorMsg != "" || e.ErrorDescription != ""
+}
+
+func (e *ErrResp) Error() string {
+	return fmt.Sprintf("ErrorCode: %d ,Error: %s ,ErrorDescription: %s ", e.ErrorCode, e.ErrorMsg, e.ErrorDescription)
 }
