@@ -1,9 +1,12 @@
 package template
 
 import (
+	"github.com/alist-org/alist/v3/internal/model"
+	"io"
+	"os"
 	"time"
 
-	"github.com/Xhofe/wopan-sdk-go"
+	"github.com/xhofe/wopan-sdk-go"
 )
 
 // do others that not defined in Driver interface
@@ -37,4 +40,27 @@ func (d *Wopan) getSpaceType() string {
 // 20230607214351
 func getTime(str string) (time.Time, error) {
 	return time.Parse("20060102150405", str)
+}
+
+// 将 FileStreamer 转换为 *os.File
+func convertToFile(stream model.FileStreamer) (*os.File, error) {
+	// 创建一个临时文件
+	tmpFile, err := os.CreateTemp("", "stream-*.tmp")
+	if err != nil {
+		return nil, err
+	}
+
+	// 将 stream 的内容复制到临时文件中
+	if _, err := io.Copy(tmpFile, stream); err != nil {
+		tmpFile.Close()
+		return nil, err
+	}
+
+	// 重新打开临时文件，以便读取
+	if _, err := tmpFile.Seek(0, io.SeekStart); err != nil {
+		tmpFile.Close()
+		return nil, err
+	}
+
+	return tmpFile, nil
 }
