@@ -2,7 +2,6 @@ package db
 
 import (
 	"fmt"
-	"sort"
 
 	"github.com/alist-org/alist/v3/internal/model"
 	"github.com/pkg/errors"
@@ -36,7 +35,7 @@ func GetStorages(pageIndex, pageSize int) ([]model.Storage, int64, error) {
 		return nil, 0, errors.Wrapf(err, "failed get storages count")
 	}
 	var storages []model.Storage
-	if err := storageDB.Order(columnName("order")).Offset((pageIndex - 1) * pageSize).Limit(pageSize).Find(&storages).Error; err != nil {
+	if err := addStorageOrder(storageDB).Order(columnName("order")).Offset((pageIndex - 1) * pageSize).Limit(pageSize).Find(&storages).Error; err != nil {
 		return nil, 0, errors.WithStack(err)
 	}
 	return storages, count, nil
@@ -63,11 +62,9 @@ func GetStorageByMountPath(mountPath string) (*model.Storage, error) {
 
 func GetEnabledStorages() ([]model.Storage, error) {
 	var storages []model.Storage
-	if err := db.Where(fmt.Sprintf("%s = ?", columnName("disabled")), false).Find(&storages).Error; err != nil {
+	err := addStorageOrder(db).Where(fmt.Sprintf("%s = ?", columnName("disabled")), false).Find(&storages).Error
+	if err != nil {
 		return nil, errors.WithStack(err)
 	}
-	sort.Slice(storages, func(i, j int) bool {
-		return storages[i].Order < storages[j].Order
-	})
 	return storages, nil
 }
