@@ -1,6 +1,7 @@
 package net
 
 import (
+	"compress/gzip"
 	"context"
 	"fmt"
 	"io"
@@ -86,9 +87,9 @@ func ServeHTTP(w http.ResponseWriter, r *http.Request, name string, modTime time
 	sendSize := size
 	var sendContent io.ReadCloser
 	ranges, err := http_range.ParseRange(rangeReq, size)
-	switch err {
-	case nil:
-	case http_range.ErrNoOverlap:
+	switch {
+	case err == nil:
+	case errors.Is(err, http_range.ErrNoOverlap):
 		if size == 0 {
 			// Some clients add a Range header to all requests to
 			// limit the size of the response. If the file is empty,
@@ -104,7 +105,7 @@ func ServeHTTP(w http.ResponseWriter, r *http.Request, name string, modTime time
 		return
 	}
 
-	if sumRangesSize(ranges) > size || size < 0 {
+	if sumRangesSize(ranges) > size {
 		// The total number of bytes in all the ranges is larger than the size of the file
 		// or unknown file size, ignore the range request.
 		ranges = nil
@@ -173,6 +174,7 @@ func ServeHTTP(w http.ResponseWriter, r *http.Request, name string, modTime time
 			pw.Close()
 		}()
 	}
+	defer sendContent.Close()
 
 	w.Header().Set("Accept-Ranges", "bytes")
 	if w.Header().Get("Content-Encoding") == "" {
@@ -191,7 +193,6 @@ func ServeHTTP(w http.ResponseWriter, r *http.Request, name string, modTime time
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 		}
 	}
-	//defer sendContent.Close()
 }
 func ProcessHeader(origin, override http.Header) http.Header {
 	result := http.Header{}
@@ -222,8 +223,19 @@ func RequestHttp(ctx context.Context, httpMethod string, headerOverride http.Hea
 	}
 	// TODO clean header with blocklist or passlist
 	res.Header.Del("set-cookie")
+	var reader io.Reader
 	if res.StatusCode >= 400 {
-		all, _ := io.ReadAll(res.Body)
+		// 根据 Content-Encoding 判断 Body 是否压缩
+		switch res.Header.Get("Content-Encoding") {
+		case "gzip":
+			// 使用gzip.NewReader解压缩
+			reader, _ = gzip.NewReader(res.Body)
+			defer reader.(*gzip.Reader).Close()
+		default:
+			// 没有Content-Encoding，直接读取
+			reader = res.Body
+		}
+		all, _ := io.ReadAll(reader)
 		_ = res.Body.Close()
 		msg := string(all)
 		log.Debugln(msg)
