@@ -14,19 +14,26 @@ import (
 
 type DownloadTask struct {
 	tache.Base
-	Url          string       `json:"url"`
-	DstDirPath   string       `json:"dst_dir_path"`
-	TempDir      string       `json:"temp_dir"`
-	DeletePolicy DeletePolicy `json:"delete_policy"`
-
-	Status            string   `json:"status"`
-	Signal            chan int `json:"-"`
-	GID               string   `json:"-"`
+	Url               string       `json:"url"`
+	DstDirPath        string       `json:"dst_dir_path"`
+	TempDir           string       `json:"temp_dir"`
+	DeletePolicy      DeletePolicy `json:"delete_policy"`
+	Toolname          string       `json:"toolname"`
+	Status            string       `json:"-"`
+	Signal            chan int     `json:"-"`
+	GID               string       `json:"-"`
 	tool              Tool
 	callStatusRetried int
 }
 
 func (t *DownloadTask) Run() error {
+	if t.tool == nil {
+		tool, err := Tools.Get(t.Toolname)
+		if err != nil {
+			return errors.WithMessage(err, "failed get tool")
+		}
+		t.tool = tool
+	}
 	if err := t.tool.Run(t); !errs.IsNotSupportError(err) {
 		if err == nil {
 			return t.Complete()
@@ -47,9 +54,7 @@ func (t *DownloadTask) Run() error {
 		return err
 	}
 	t.GID = gid
-	var (
-		ok bool
-	)
+	var ok bool
 outer:
 	for {
 		select {
@@ -70,6 +75,18 @@ outer:
 	}
 	if err != nil {
 		return err
+	}
+	if t.tool.Name() == "pikpak" {
+		return nil
+	}
+	if t.tool.Name() == "115 Cloud" {
+		// hack for 115
+		<-time.After(time.Second * 1)
+		err := t.tool.Remove(t)
+		if err != nil {
+			log.Errorln(err.Error())
+		}
+		return nil
 	}
 	t.Status = "offline download completed, maybe transferring"
 	// hack for qBittorrent
@@ -123,6 +140,12 @@ func (t *DownloadTask) Complete() error {
 		files []File
 		err   error
 	)
+	if t.tool.Name() == "pikpak" {
+		return nil
+	}
+	if t.tool.Name() == "115 Cloud" {
+		return nil
+	}
 	if getFileser, ok := t.tool.(GetFileser); ok {
 		files = getFileser.GetFiles(t)
 	} else {
@@ -132,13 +155,14 @@ func (t *DownloadTask) Complete() error {
 		}
 	}
 	// upload files
-	for i, _ := range files {
+	for i := range files {
 		file := files[i]
 		TransferTaskManager.Add(&TransferTask{
 			file:         file,
-			dstDirPath:   t.DstDirPath,
-			tempDir:      t.TempDir,
-			deletePolicy: t.DeletePolicy,
+			DstDirPath:   t.DstDirPath,
+			TempDir:      t.TempDir,
+			DeletePolicy: t.DeletePolicy,
+			FileDir:      file.Path,
 		})
 	}
 	return nil
@@ -152,6 +176,4 @@ func (t *DownloadTask) GetStatus() string {
 	return t.Status
 }
 
-var (
-	DownloadTaskManager *tache.Manager[*DownloadTask]
-)
+var DownloadTaskManager *tache.Manager[*DownloadTask]
